got fetch to work, working on add now
