+++ smurfs/src/components/Smurfs.js
@@ -1,14 +1,35 @@
 import React, { useEffect } from "react";
+import { connect } from "react-redux";
+
+import Smurf from "./Smurf"
+
+import { fetchSmurf } from "../actions";
 
 import PropTypes from "prop-types";
 
-function Smurfs({ isFetching }) {
+function Smurfs({ fetchSmurf, data }) {
   useEffect(() => {
-    console.log(isFetching);
+    fetchSmurf();
   }, []);
-  return <div />;
+  return <div>
+    {data.map(smurf => <Smurf key={smurf.id} smurf={smurf}/>)}
+  </div>;
 }
 
 Smurfs.propTypes = {};
 
-export default Smurfs;
+const mapStateToProps = state => {
+  return {
+    data: state.data,
+    error: state.error,
+    isFetching: state.isFetching,
+    isAdding: state.isAdding,
+    isUpdating: state.isUpdating,
+    isDeleting: state.isDeleting,
+  };
+};
+
+export default connect(
+  mapStateToProps,
+  { fetchSmurf },
+)(Smurfs);
